requirements.txt
python-telegram-bot==20.7
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.2.0
python-docx==1.1.0
anthropic==0.18.0
openai==1.12.0
reportlab==4.1.0
openpyxl==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.1
PyMuPDF==1.23.26
